pyproject.toml
[tool.poetry]
name = "hw"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"


[tool.poetry.group.dev.dependencies]
requests = "^2.32.3"
pytest-cov = "^5.0.0"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.0"
black = "^24.4.2"
isort = "^5.13.2"
mypy = "^1.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
exclude = '\.git'

[tool.isort]
profile = "black"
line_length = 119
force_single_line = true

[tool.mypy]
python_version = 3.8
ignore_missing_imports = true
disallow_untyped_defs = true
warn_return_any = true
exclude = 'venv'
